pyproject.toml
[tool.poetry]
name = "flask-jwt-authentication"
version = "0.1.0"
description = ""
authors = ["Demos Petsas <demos@petsas.consulting>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.3"
Flask-RESTful = "^0.3.9"
python-dotenv = "^1.0.0"
Flask-SQLAlchemy = "^3.0.3"
Flask-Migrate = "^4.0.4"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.29.0"
Flask-Cors = "^3.0.10"
Flask-JWT-Extended = "^4.4.4"
passlib = "^1.7.4"
flask-query-builder = "^0.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
